Beginning of Draggable Form Icons representing the Object with an embedded Record Id via a RecordSet Plus a lot of warning clean up.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@838 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/conversion/AttrUtils.java
@@ -105,7 +105,7 @@ public class AttrUtils
                 {
                     HibernateUtil.beginTransaction();
                     AttributeDef attrsDef = new AttributeDef();
-                    attrsDef.setTableType((short)tableType.getType());
+                    attrsDef.setTableType(tableType.getType());
                     attrsDef.setCollectionObjDef(colObjDef);
                     attrsDef.setPrepType(prepType);
                     attrsDef.setFieldName(attrNames[i]);
